+++ template-parts/components/components.php
@@ -4,12 +4,12 @@ $field = $args['field'] ?? 'components';
 
 //preint_r($field);
 
-if ($field) {
+if (!empty($field)) {
   foreach ($field as $layout) {
     $acf_fc_layout = $layout['acf_fc_layout'];
     if ($acf_fc_layout) {
       $template = 'template-parts/components/' . $acf_fc_layout;
-      echo '<div class="my-8">';
+      echo '<div class="component-wrapper">';
       get_template_part($template, '', array('field' => $layout));
       echo '</div>';
     }